+++ package.json
@@ -5,7 +5,7 @@
   "author": "Harper",
   "license": "MIT",
   "type": "module",
-  "exports": "./dist/index.js",
+  "exports": "./dist/extended-iterable.js",
   "types": "./dist/index.d.ts",
   "files": [
     "dist"
