+++ src/extended-iterable.ts
